fixup! refactor: port DiffviewOPen command
