+++ lua/diffview/scene/views/diff/diff_view.lua
@@ -13,7 +13,7 @@ local RevType = lazy.access("diffview.vcs.rev", "RevType") ---@type RevType|Lazy
 local StandardView = lazy.access("diffview.scene.views.standard.standard_view", "StandardView") ---@type StandardView|LazyModule
 local async = lazy.require("plenary.async") ---@module "plenary.async"
 local debounce = lazy.require("diffview.debounce") ---@module "diffview.debounce"
-local vcs = lazy.require("diffview.vcs") ---@module "diffview.vcs"
+local vcs = lazy.require("diffview.vcs.utils") ---@module "diffview.vcs"
 local logger = lazy.require("diffview.logger") ---@module "diffview.logger"
 local utils = lazy.require("diffview.utils") ---@module "diffview.utils"
 local config = lazy.require("diffview.config") ---@module "diffview.config"
