+++ lua/diffview/vcs/adapters/git/utils.lua
@@ -15,281 +15,6 @@ local api = vim.api
 
 local M = {}
 
----@param ctx GitContext
----@param left Rev
----@param right Rev
----@param args string[]
----@param kind git.FileKind
----@param opt git.utils.LayoutOpt
----@param callback function
-local tracked_files = async.wrap(function(ctx, left, right, args, kind, opt, callback)
-  ---@type FileEntry[]
-  local files = {}
-  ---@type FileEntry[]
-  local conflicts = {}
-  ---@type CountDownLatch
-  local latch = CountDownLatch(2)
-  local debug_opt = {
-    context = "git.utils>tracked_files()",
-    func = "s_debug",
-    debug_level = 1,
-    no_stdout = true,
-  }
-
-  ---@param job Job
-  local function on_exit(job)
-    utils.handle_job(job, { debug_opt = debug_opt })
-    latch:count_down()
-  end
-
-  local namestat_job = Job:new({
-    command = git_bin(),
-    args = utils.vec_join(git_args(), "diff", "--ignore-submodules", "--name-status", args),
-    cwd = ctx.toplevel,
-    on_exit = on_exit
-  })
-  local numstat_job = Job:new({
-    command = git_bin(),
-    args = utils.vec_join(git_args(), "diff", "--ignore-submodules", "--numstat", args),
-    cwd = ctx.toplevel,
-    on_exit = on_exit
-  })
-
-  namestat_job:start()
-  numstat_job:start()
-  latch:await()
-  local out_status
-  if not (#namestat_job:result() == #numstat_job:result()) then
-    out_status = ensure_output(2, { namestat_job, numstat_job }, "git.utils>tracked_files()")
-  end
-
-  if out_status == JobStatus.ERROR or not (namestat_job.code == 0 and numstat_job.code == 0) then
-    callback(utils.vec_join(namestat_job:stderr_result(), numstat_job:stderr_result()), nil)
-    return
-  end
-
-  local numstat_out = numstat_job:result()
-  local data = {}
-  local conflict_map = {}
-
-  for i, s in ipairs(namestat_job:result()) do
-    local status = s:sub(1, 1):gsub("%s", " ")
-    local name = s:match("[%a%s][^%s]*\t(.*)")
-    local oldname
-
-    if name:match("\t") ~= nil then
-      oldname = name:match("(.*)\t")
-      name = name:gsub("^.*\t", "")
-    end
-
-    local stats = {
-      additions = tonumber(numstat_out[i]:match("^%d+")),
-      deletions = tonumber(numstat_out[i]:match("^%d+%s+(%d+)")),
-    }
-
-    if not stats.additions or not stats.deletions then
-      stats = nil
-    end
-
-    if not (status == "U" and kind == "staged") then
-      table.insert(data, {
-        status = status,
-        name = name,
-        oldname = oldname,
-        stats = stats,
-      })
-    end
-
-    if status == "U" then
-      conflict_map[name] = data[#data]
-    end
-  end
-
-  if kind == "working" and next(conflict_map) then
-    data = vim.tbl_filter(function(v)
-      return not conflict_map[v.name]
-    end, data)
-
-    for _, v in pairs(conflict_map) do
-      table.insert(conflicts, FileEntry.with_layout(opt.merge_layout, {
-        git_ctx = ctx,
-        path = v.name,
-        oldpath = v.oldname,
-        status = "U",
-        kind = "conflicting",
-        rev_ours = Rev(RevType.STAGE, 2),
-        rev_main = Rev(RevType.LOCAL),
-        rev_theirs = Rev(RevType.STAGE, 3),
-        rev_base = Rev(RevType.STAGE, 1),
-      }))
-    end
-  end
-
-  for _, v in ipairs(data) do
-    table.insert(files, FileEntry.for_d2(opt.default_layout, {
-      git_ctx = ctx,
-      path = v.name,
-      oldpath = v.oldname,
-      status = v.status,
-      stats = v.stats,
-      kind = kind,
-      rev_a = left,
-      rev_b = right,
-    }))
-  end
-
-  callback(nil, files, conflicts)
-end, 7)
-
----@param ctx GitContext
----@param left Rev
----@param right Rev
----@param opt git.utils.LayoutOpt
----@param callback function
-local untracked_files = async.wrap(function(ctx, left, right, opt, callback)
-  Job:new({
-    command = git_bin(),
-    args = utils.vec_join(git_args(), "ls-files", "--others", "--exclude-standard" ),
-    cwd = ctx.toplevel,
-    ---@type Job
-    on_exit = function(j)
-      utils.handle_job(j, {
-        debug_opt = {
-          context = "git.utils>untracked_files()",
-          func = "s_debug",
-          debug_level = 1,
-          no_stdout = true,
-        }
-      })
-
-      if j.code ~= 0 then
-        callback(j:stderr_result() or {}, nil)
-        return
-      end
-
-      local files = {}
-      for _, s in ipairs(j:result()) do
-        table.insert(files, FileEntry.for_d2(opt.default_layout, {
-          git_ctx = ctx,
-          path = s,
-          status = "?",
-          kind = "working",
-          rev_a = left,
-          rev_b = right,
-        }))
-      end
-      callback(nil, files)
-    end
-  }):start()
-end, 5)
-
----Get a list of files modified between two revs.
----@param ctx GitContext
----@param left Rev
----@param right Rev
----@param path_args string[]|nil
----@param dv_opt DiffViewOptions
----@param opt git.utils.LayoutOpt
----@param callback function
----@return string[]? err
----@return FileDict?
-M.diff_file_list = async.wrap(function(ctx, left, right, path_args, dv_opt, opt, callback)
-  ---@type FileDict
-  local files = FileDict()
-  ---@type CountDownLatch
-  local latch = CountDownLatch(2)
-  local rev_args = M.rev_to_args(left, right)
-  local errors = {}
-
-  tracked_files(
-    ctx,
-    left,
-    right,
-    utils.vec_join(
-      rev_args,
-      "--",
-      path_args
-    ),
-    "working",
-    opt,
-    function (err, tfiles, tconflicts)
-      if err then
-        errors[#errors+1] = err
-        utils.err("Failed to get git status for tracked files!", true)
-        latch:count_down()
-        return
-      end
-
-      files:set_working(tfiles)
-      files:set_conflicting(tconflicts)
-      local show_untracked = dv_opt.show_untracked
-
-      if show_untracked == nil then
-        show_untracked = M.show_untracked(ctx.toplevel)
-      end
-
-      if not (show_untracked and M.has_local(left, right)) then
-        latch:count_down()
-        return
-      end
-
-      ---@diagnostic disable-next-line: redefined-local
-      local err, ufiles = untracked_files(ctx, left, right, opt)
-      if err then
-        errors[#errors+1] = err
-        utils.err("Failed to get git status for untracked files!", true)
-        latch:count_down()
-      else
-        files:set_working(utils.vec_join(files.working, ufiles))
-
-        utils.merge_sort(files.working, function(a, b)
-          return a.path:lower() < b.path:lower()
-        end)
-        latch:count_down()
-      end
-    end
-  )
-
-  if not (left.type == RevType.STAGE and right.type == RevType.LOCAL) then
-    latch:count_down()
-  else
-    local left_rev = M.head_rev(ctx.toplevel) or Rev.new_null_tree()
-    local right_rev = Rev(RevType.STAGE, 0)
-    tracked_files(
-      ctx,
-      left_rev,
-      right_rev,
-      utils.vec_join(
-        "--cached",
-        left_rev.commit,
-        "--",
-        path_args
-      ),
-      "staged",
-      opt,
-      function(err, tfiles)
-        if err then
-          errors[#errors+1] = err
-          utils.err("Failed to get git status for staged files!", true)
-          latch:count_down()
-          return
-        end
-        files:set_staged(tfiles)
-        latch:count_down()
-      end
-    )
-  end
-
-  latch:await()
-  if #errors > 0 then
-    callback(utils.vec_join(unpack(errors)), nil)
-    return
-  end
-
-  files:update_file_trees()
-  callback(nil, files)
-end, 7)
-
 ---@class git.utils.PreparedLogOpts
 ---@field rev_range string
 ---@field base Rev
@@ -805,26 +530,6 @@ local function parse_fh_line_trace_data(state)
   return true
 end
 
----Convert revs to git rev args.
----@param left Rev
----@param right Rev
----@return string[]
-function M.rev_to_args(left, right)
-  assert(
-    not (left.type == RevType.LOCAL and right.type == RevType.LOCAL),
-    "Can't diff LOCAL against LOCAL!"
-  )
-
-  if left.type == RevType.COMMIT and right.type == RevType.COMMIT then
-    return { left.commit .. ".." .. right.commit }
-  elseif left.type == RevType.STAGE and right.type == RevType.LOCAL then
-    return {}
-  elseif left.type == RevType.COMMIT and right.type == RevType.STAGE then
-    return { "--cached", left.commit }
-  else
-    return { left.commit }
-  end
-end
 
 ---Convert revs to string representation.
 ---@param left Rev
@@ -1034,14 +739,6 @@ function M.pathspec_modify(pathspec, mods)
   return magic .. utils.path:vim_fnamemodify(pattern, mods)
 end
 
----Check if any of the given revs are LOCAL.
----@param left Rev
----@param right Rev
----@return boolean
-function M.has_local(left, right)
-  return left.type == RevType.LOCAL or right.type == RevType.LOCAL
-end
-
 ---Strange trick to check if a file is binary using only git.
 ---@param toplevel string
 ---@param path string
@@ -1065,17 +762,6 @@ function M.is_binary(toplevel, path, rev)
 
   local _, code = M.exec_sync(cmd, { cwd = toplevel, silent = true })
   return code ~= 0
-end
-
----Check if status for untracked files is disabled for a given git repo.
----@param toplevel string
----@return boolean
-function M.show_untracked(toplevel)
-  local out = M.exec_sync(
-    { "config", "status.showUntrackedFiles" },
-    { cwd = toplevel, silent = true }
-  )
-  return vim.trim(out[1] or "") ~= "no"
 end
 
 ---Get the diff status letter for a file for a given rev.
