+++ lua/diffview/vcs/utils.lua
@@ -1,6 +1,8 @@
+local CountDownLatch = require("diffview.control").CountDownLatch
 local utils = require("diffview.utils")
 local async = require("plenary.async")
 local logger = require("diffview.logger")
+local FileDict = require("diffview.vcs.file_dict").FileDict
 local Job = require("plenary.job")
 local Semaphore = require('diffview.control').Semaphore
 
@@ -145,6 +147,281 @@ M.show = async.wrap(function(adapter, args, callback)
   queue_sync_job(job)
 
 end, 3)
+
+---@param adapter VCSAdapter
+---@param left Rev
+---@param right Rev
+---@param args string[]
+---@param kind git.FileKind
+---@param opt git.utils.LayoutOpt
+---@param callback function
+local tracked_files = async.wrap(function(adapter, left, right, args, kind, opt, callback)
+  ---@type FileEntry[]
+  local files = {}
+  ---@type FileEntry[]
+  local conflicts = {}
+  ---@type CountDownLatch
+  local latch = CountDownLatch(2)
+  local debug_opt = {
+    context = "git.utils>tracked_files()",
+    func = "s_debug",
+    debug_level = 1,
+    no_stdout = true,
+  }
+
+  ---@param job Job
+  local function on_exit(job)
+    utils.handle_job(job, { debug_opt = debug_opt })
+    latch:count_down()
+  end
+
+  local namestat_job = Job:new({
+    command = adapter:bin(),
+    args = adapter:get_namestat_args(args),
+    cwd = adapter.ctx.toplevel,
+    on_exit = on_exit
+  })
+  local numstat_job = Job:new({
+    command = adapter:bin(),
+    args = adapter:get_numstat_args(args),
+    cwd = adapter.ctx.toplevel,
+    on_exit = on_exit
+  })
+
+  namestat_job:start()
+  numstat_job:start()
+  latch:await()
+  local out_status
+  if not (#namestat_job:result() == #numstat_job:result()) then
+    out_status = ensure_output(2, { namestat_job, numstat_job }, "git.utils>tracked_files()")
+  end
+
+  if out_status == JobStatus.ERROR or not (namestat_job.code == 0 and numstat_job.code == 0) then
+    callback(utils.vec_join(namestat_job:stderr_result(), numstat_job:stderr_result()), nil)
+    return
+  end
+
+  local numstat_out = numstat_job:result()
+  local data = {}
+  local conflict_map = {}
+
+  for i, s in ipairs(namestat_job:result()) do
+    local status = s:sub(1, 1):gsub("%s", " ")
+    local name = s:match("[%a%s][^%s]*\t(.*)")
+    local oldname
+
+    if name:match("\t") ~= nil then
+      oldname = name:match("(.*)\t")
+      name = name:gsub("^.*\t", "")
+    end
+
+    local stats = {
+      additions = tonumber(numstat_out[i]:match("^%d+")),
+      deletions = tonumber(numstat_out[i]:match("^%d+%s+(%d+)")),
+    }
+
+    if not stats.additions or not stats.deletions then
+      stats = nil
+    end
+
+    if not (status == "U" and kind == "staged") then
+      table.insert(data, {
+        status = status,
+        name = name,
+        oldname = oldname,
+        stats = stats,
+      })
+    end
+
+    if status == "U" then
+      conflict_map[name] = data[#data]
+    end
+  end
+
+  if kind == "working" and next(conflict_map) then
+    data = vim.tbl_filter(function(v)
+      return not conflict_map[v.name]
+    end, data)
+
+    for _, v in pairs(conflict_map) do
+      table.insert(conflicts, FileEntry.with_layout(opt.merge_layout, {
+        git_ctx = adapter,
+        path = v.name,
+        oldpath = v.oldname,
+        status = "U",
+        kind = "conflicting",
+        rev_ours = Rev(RevType.STAGE, 2),
+        rev_main = Rev(RevType.LOCAL),
+        rev_theirs = Rev(RevType.STAGE, 3),
+        rev_base = Rev(RevType.STAGE, 1),
+      }))
+    end
+  end
+
+  for _, v in ipairs(data) do
+    table.insert(files, FileEntry.for_d2(opt.default_layout, {
+      git_ctx = adapter,
+      path = v.name,
+      oldpath = v.oldname,
+      status = v.status,
+      stats = v.stats,
+      kind = kind,
+      rev_a = left,
+      rev_b = right,
+    }))
+  end
+
+  callback(nil, files, conflicts)
+end, 7)
+
+---@param adapter VCSAdapter
+---@param left Rev
+---@param right Rev
+---@param opt git.utils.LayoutOpt
+---@param callback function
+local untracked_files = async.wrap(function(adapter, left, right, opt, callback)
+  Job:new({
+    command = adapter:bin(),
+    args = adapter:get_files_args(),
+    cwd = adapter.ctx.toplevel,
+    ---@type Job
+    on_exit = function(j)
+      utils.handle_job(j, {
+        debug_opt = {
+          context = "git.utils>untracked_files()",
+          func = "s_debug",
+          debug_level = 1,
+          no_stdout = true,
+        }
+      })
+
+      if j.code ~= 0 then
+        callback(j:stderr_result() or {}, nil)
+        return
+      end
+
+      local files = {}
+      for _, s in ipairs(j:result()) do
+        table.insert(files, FileEntry.for_d2(opt.default_layout, {
+          git_ctx = adapter,
+          path = s,
+          status = "?",
+          kind = "working",
+          rev_a = left,
+          rev_b = right,
+        }))
+      end
+      callback(nil, files)
+    end
+  }):start()
+end, 5)
+
+---Get a list of files modified between two revs.
+---@param adapter VCSAdapter
+---@param left Rev
+---@param right Rev
+---@param path_args string[]|nil
+---@param dv_opt DiffViewOptions
+---@param opt git.utils.LayoutOpt
+---@param callback function
+---@return string[]? err
+---@return FileDict?
+M.diff_file_list = async.wrap(function(adapter, left, right, path_args, dv_opt, opt, callback)
+  ---@type FileDict
+  local files = FileDict()
+  ---@type CountDownLatch
+  local latch = CountDownLatch(2)
+  local rev_args = adapter:rev_to_args(left, right)
+  local errors = {}
+
+  tracked_files(
+    adapter,
+    left,
+    right,
+    utils.vec_join(
+      rev_args,
+      "--",
+      path_args
+    ),
+    "working",
+    opt,
+    function (err, tfiles, tconflicts)
+      if err then
+        errors[#errors+1] = err
+        utils.err("Failed to get git status for tracked files!", true)
+        latch:count_down()
+        return
+      end
+
+      files:set_working(tfiles)
+      files:set_conflicting(tconflicts)
+      local show_untracked = dv_opt.show_untracked
+
+      if show_untracked == nil then
+        show_untracked = adapter:show_untracked()
+      end
+
+      if not (show_untracked and adapter:has_local(left, right)) then
+        latch:count_down()
+        return
+      end
+
+      ---@diagnostic disable-next-line: redefined-local
+      local err, ufiles = untracked_files(adapter, left, right, opt)
+      if err then
+        errors[#errors+1] = err
+        utils.err("Failed to get git status for untracked files!", true)
+        latch:count_down()
+      else
+        files:set_working(utils.vec_join(files.working, ufiles))
+
+        utils.merge_sort(files.working, function(a, b)
+          return a.path:lower() < b.path:lower()
+        end)
+        latch:count_down()
+      end
+    end
+  )
+
+  if not (left.type == RevType.STAGE and right.type == RevType.LOCAL) then
+    latch:count_down()
+  else
+    local left_rev = M.head_rev(ctx.toplevel) or Rev.new_null_tree()
+    local right_rev = Rev(RevType.STAGE, 0)
+    tracked_files(
+      adapter,
+      left_rev,
+      right_rev,
+      utils.vec_join(
+        "--cached",
+        left_rev.commit,
+        "--",
+        path_args
+      ),
+      "staged",
+      opt,
+      function(err, tfiles)
+        if err then
+          errors[#errors+1] = err
+          utils.err("Failed to get git status for staged files!", true)
+          latch:count_down()
+          return
+        end
+        files:set_staged(tfiles)
+        latch:count_down()
+      end
+    )
+  end
+
+  latch:await()
+  if #errors > 0 then
+    callback(utils.vec_join(unpack(errors)), nil)
+    return
+  end
+
+  files:update_file_trees()
+  callback(nil, files)
+end, 7)
 
 
 ---@param arg_lead string
